feat: [vault] vault functional move to server plugin.
1. as title.
2. function include transparent encryption and decryption, lock screen lock,
auto locking, password input protection.
