+++ src/plugins/server/CMakeLists.txt
@@ -3,3 +3,4 @@ cmake_minimum_required(VERSION 3.10)
 add_subdirectory(core/serverplugin-core)
 add_subdirectory(serverplugin-filemanager1)
 add_subdirectory(serverplugin-grandsearchdaemon)
+add_subdirectory(serverplugin-vaultdaemon)
