+++ src/plugins/server/serverplugin-vaultdaemon/vaultdaemon.json
@@ -1,0 +1,7 @@
+{
+    "Name" : "serverplugin-vaultdaemon",
+    "Version" : "1.0.0",
+    "Description" : "The vault helper plugin for the dde-file-manager-server.",
+    "Depends" : [
+    ]
+}
